move router_service to a subfolder
