+++ apollo-router/src/orbiter/mod.rs
@@ -23,7 +23,7 @@ use crate::executable::Opt;
 use crate::plugin::DynPlugin;
 use crate::router_factory::RouterSuperServiceFactory;
 use crate::router_factory::YamlRouterFactory;
-use crate::services::router_service::RouterCreator;
+use crate::services::router::service::RouterCreator;
 use crate::services::HasSchema;
 use crate::spec::Schema;
 use crate::Configuration;
