+++ apollo-router/src/plugins/include_subgraph_errors.rs
@@ -95,7 +95,7 @@ mod test {
     use crate::services::layers::persisted_queries::PersistedQueryLayer;
     use crate::services::layers::query_analysis::QueryAnalysisLayer;
     use crate::services::router;
-    use crate::services::router_service::RouterCreator;
+    use crate::services::router::service::RouterCreator;
     use crate::services::HasSchema;
     use crate::services::PluggableSupergraphServiceBuilder;
     use crate::services::SupergraphRequest;
