+++ apollo-router/src/services/layers/apq.rs
@@ -206,9 +206,9 @@ mod apq_tests {
     use super::*;
     use crate::error::Error;
     use crate::graphql::Response;
+    use crate::services::router::service::from_supergraph_mock_callback;
+    use crate::services::router::service::from_supergraph_mock_callback_and_configuration;
     use crate::services::router::ClientRequestAccepts;
-    use crate::services::router_service::from_supergraph_mock_callback;
-    use crate::services::router_service::from_supergraph_mock_callback_and_configuration;
     use crate::Configuration;
     use crate::Context;
