+++ apollo-router/src/services/layers/content_negotiation.rs
@@ -22,9 +22,9 @@ use crate::graphql;
 use crate::layers::sync_checkpoint::CheckpointService;
 use crate::layers::ServiceExt as _;
 use crate::services::router;
+use crate::services::router::service::MULTIPART_DEFER_HEADER_VALUE;
+use crate::services::router::service::MULTIPART_SUBSCRIPTION_HEADER_VALUE;
 use crate::services::router::ClientRequestAccepts;
-use crate::services::router_service::MULTIPART_DEFER_HEADER_VALUE;
-use crate::services::router_service::MULTIPART_SUBSCRIPTION_HEADER_VALUE;
 use crate::services::supergraph;
 use crate::services::APPLICATION_JSON_HEADER_VALUE;
 use crate::services::MULTIPART_DEFER_CONTENT_TYPE;
