+++ apollo-router/src/services/mod.rs
@@ -29,7 +29,6 @@ pub(crate) mod layers;
 pub(crate) mod new_service;
 pub(crate) mod query_planner;
 pub mod router;
-pub(crate) mod router_service;
 pub mod subgraph;
 pub(crate) mod subgraph_service;
 pub mod supergraph;
