+++ apollo-router/src/services/router.rs
@@ -17,8 +17,6 @@ use serde_json_bytes::Value;
 use static_assertions::assert_impl_all;
 use tower::BoxError;
 
-use super::router_service::MULTIPART_DEFER_HEADER_VALUE;
-use super::router_service::MULTIPART_SUBSCRIPTION_HEADER_VALUE;
 use super::supergraph;
 use crate::graphql;
 use crate::http_ext::header_map;
@@ -26,12 +24,16 @@ use crate::json_ext::Path;
 use crate::services::TryIntoHeaderName;
 use crate::services::TryIntoHeaderValue;
 use crate::Context;
+use service::MULTIPART_DEFER_HEADER_VALUE;
+use service::MULTIPART_SUBSCRIPTION_HEADER_VALUE;
 
 pub type BoxService = tower::util::BoxService<Request, Response, BoxError>;
 pub type BoxCloneService = tower::util::BoxCloneService<Request, Response, BoxError>;
 pub type ServiceResult = Result<Response, BoxError>;
 pub type Body = hyper::Body;
 pub type Error = hyper::Error;
+
+pub(crate) mod service;
 
 assert_impl_all!(Request: Send);
 /// Represents the router processing step of the processing pipeline.
